+++ buildSrc/src/main/resources/META-INF/gradle-plugins/elasticsearch.extra-clusters-test.properties
@@ -17,4 +17,4 @@
 # under the License.
 #
 
-implementation-class=org.elasticsearch.gradle.test.MultiClusterTestPlugin
+implementation-class=org.elasticsearch.gradle.test.ExtraClustersTestPlugin
